+++ source/Irrlicht/CImageLoaderPCX.cpp
@@ -1,4 +1,4 @@
-// Copyright (C) 2002-2011 Nikolaus Gebhardt
+// Copyright (C) 2002-2012 Nikolaus Gebhardt
 // This file is part of the "Irrlicht Engine".
 // For conditions of distribution and use, see copyright notice in irrlicht.h
 
@@ -141,7 +141,7 @@ IImage* CImageLoaderPCX::loadImage(io::IReadFile* file) const
 			memset(PCXData+offset, value, cnt);
 		else
 		{
-			for (u32 i=0; i<cnt; ++i)
+			for (u8 i=0; i<cnt; ++i)
 			{
 				PCXData[linestart+lineoffset]=value;
 				lineoffset += 3;
